Add old logic

+++ Models/NewPolicyModel.cs
@@ -10,7 +10,7 @@ namespace Models
         public string LastName { get; set; }
         public bool HasDriverLicence { get; set; }
         public bool HasChildren { get; set; }
-        public int AnnualMilage { get; set; }
+        public decimal AnnualMilage { get; set; }
         public bool HasCars { get; set; }
         public int DriverLicenceSince { get; set; }
     }
